requirements.txt
pytest==6.2.5
python-dotenv==0.19.0
# Add other dependencies as needed

